requirements.txt
fastapi
uvicorn
websockets
sqlalchemy
pygame-ce
pygame-gui

pyqt6-sip